# Optimized BclHelpers.WriteDecimal