Optimized BclHelpers.WriteDecimal by accessing the inner fields of the decimal using a custom struct with field offsets. This avoids the int[] allocation in decimal.GetBits() and is also faster:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.16299.611 (1709/FallCreatorsUpdate/Redstone3)
Intel Xeon CPU E5-1620 v2 3.70GHz, 1 CPU, 8 logical and 4 physical cores
Frequency=3613277 Hz, Resolution=276.7571 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3132.0
  DefaultJob : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3132.0


           Method |      Mean |     Error |    StdDev | Scaled |  Gen 0 | Allocated |
----------------- |----------:|----------:|----------:|-------:|-------:|----------:|
          GetBits | 10.227 ns | 0.0427 ns | 0.0399 ns |   1.00 | 0.0076 |      40 B |
 StructConversion |  3.048 ns | 0.0091 ns | 0.0081 ns |   0.30 |      - |       0 B |
```
Benchmark Code: https://gist.github.com/szehetner/dab144926f1f1e1648ee342e47bf86d4