@mgravell 2019-05-08
First, sorry for the delay - I got tied up with things!

I really like the idea of this, and I see what it is doing. Nice approach. I'm a little cautious, however, that this depends on the internal `decimal` layout in ways that will be really hard for us to be absolutely sure about in a x-plat way. Maybe I'm being paranoid, or maybe I need to think more.

Definitely interested in this, I'm just... cautious. Let me ponder on it a bit (more)