@mgravell 2019-05-08
I think what we need here is a sentinel test and fallback, so:

```
static readonly book s_optimized = VerifyLayout();
```

And check the bool in the code, using the fast version or old code respectively. The JIT now recognizes this pattern and can remove the test from the runtime code, at least on new .NET Core

The verification sentinel should be a known value that tests all the fields with no zero bytes (unless they are required to be zero), and no duplicates. If the values are as expected: we're all golden.

Thoughts?