@mgravell 2019-05-09
Manually merged with the extra check added - seems good, thanks!