@szehetner 2019-05-09
Thanks, yes the extra check makes sense, even though the layout is unlikely to change.

BTW, I think the same pattern could be used for WriteGuid to avoid the Guid.ToByteArray() call, but I haven't looked at that in detail.