@mgravell 2019-05-09
I agree that the layout is **extremely** unlikely to ever be wrong, but: my head hurts too much at the idea of checking each combination of CPU-endianness of different OSes with each framework/runtime. So it is easier to add the test :)

Good spot on `Guid` - I'll hack that in a bit quick...