@mgravell 2019-05-09
Here's the guid code, if you're interested: https://github.com/mgravell/protobuf-net/commit/6e57ce00deedc8db4e4196a37ae2a675a112e913