@mgravell 2019-05-09
actually, made it a *lot* simpler now; it may turn out to be non-optimized on exotic CPUs, but: I'm OK with this