@szehetner 2019-05-09
Yes, looks good for Guids now as well.

Just out of interest, how does this JIT optimization work? Does it mean that if the JIT sees a `static readonly _field = Init()` it can assume that because the field is readonly it won't change any more so it can treat it like a constant and in methods jitted after the field was initialized it can eliminate branches depending on it? 
I knew the JIT could eliminate code based on constants and some generic type checks, but not based on the result of methods, i.e. information only known at runtime. If it can do that now, that opens up a few interesting optimization possibilities.