@mgravell 2019-05-09
Yes, this is now a thing in .NET Core 3 - it used to just impact things that mapped to code implementations like BitConverter.IsLittleEndian, Vector.IsHardwareAccelerated, but now the JIT does this for readonly static fields too, as long as they are initialized in the right way, which comes down to the `beforefieldinit` flag, as I understand it. As a consequence:

- static field initializer: good
- static constructor: bad

This ties into a related change where you now **cannot** use reflection to change the value of a static readonly field, as there's a good chance that the running code won't see it - depending on whether the JIT has already run.

IIRC it can also do some de-virtualizing of calls, based on using the runtime type of an object rather than the declared type - I haven't played with this, but things like:

``` c#
static readonly SomeBaseType s_whatever = new SomeConcreteType();
...
s_whatever.SomeVirtualMethod();
```

this should now be able to devirtualize to `SomeConcreteType.SomeVirtualMethod` as though it used `call` instead of `callvirt`.

Make sense?