@mgravell 2019-05-09
here's one half of this (the devirtualization stuff): https://github.com/dotnet/coreclr/pull/20886